Move namespace filter to separate module
